refactor(services): migrate status_query_service to shared/status_query
- Move status_query_service.py → shared/status_query.py
- Update lazy import mappings and TYPE_CHECKING imports in services/__init__.py
- Update imports in orchestra/status.py, task/resume.py, flow/classifier.py, task/status.py

Part of issue #2586: Phase 3-cleanup
