+++ src/vibe3/services/shared/status_query.py
