+++ src/vibe3/services/task/resume.py
@@ -16,7 +16,7 @@ from vibe3.models import IssueState
 from vibe3.services.flow.service import FlowService
 from vibe3.services.issue.flow import IssueFlowService
 from vibe3.services.shared.label_service import LabelService
-from vibe3.services.status_query_service import StatusQueryService
+from vibe3.services.shared.status_query import StatusQueryService
 
 if TYPE_CHECKING:
     from vibe3.models import FlowStatusResponse
