+++ src/vibe3/services/task/status.py
@@ -4,17 +4,17 @@ from dataclasses import dataclass
 from typing import Any, cast
 
 from vibe3.models import FlowStatusResponse, IssueState, OrchestraConfig
+from vibe3.services import (
+    StatusQueryService,
+    is_auto_task_branch,
+    is_dev_collab_branch,
+)
 from vibe3.services.flow.service import FlowService
 from vibe3.services.orchestra.status import OrchestraSnapshot
 from vibe3.services.shared.labels import (
     has_orchestra_governed,
     has_roadmap_label,
     normalize_labels,
-)
-from vibe3.services.status_query_service import (
-    StatusQueryService,
-    is_auto_task_branch,
-    is_dev_collab_branch,
 )
 from vibe3.services.task.classifier import TaskStatusBucket, classify_task_status
